backend/requirements.txt
# FastAPI and Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Data validation and serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Weaviate client
weaviate-client>=4.16.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# CORS
fastapi-cors>=0.0.6

# Additional utilities
python-dateutil>=2.8.2

sqlalchemy>=2.0.0

requests>=2.31.0